conversion of multiply expressions into numeric values is now explicit

+++ tests/expressions.cpp
@@ -1812,7 +1812,7 @@ BOOST_AUTO_TEST_CASE( dot )
 {
   // Test the dot expression function
   int result = 0;
-  BOOST_REQUIRE_NO_THROW(result = a("a,b,c") * b("a,b,c") );
+  BOOST_REQUIRE_NO_THROW(result = static_cast<int>(a("a,b,c") * b("a,b,c")) );
   BOOST_REQUIRE_NO_THROW(result += a("a,b,c") * b("a,b,c") );
   BOOST_REQUIRE_NO_THROW(result -= a("a,b,c") * b("a,b,c") );
   BOOST_REQUIRE_NO_THROW(result *= a("a,b,c") * b("a,b,c") );
@@ -1837,7 +1837,7 @@ BOOST_AUTO_TEST_CASE( dot_permute )
   Permutation perm({2, 1, 0});
   // Test the dot expression function
   int result = 0;
-  BOOST_REQUIRE_NO_THROW(result = a("a,b,c") * b("c,b,a") );
+  BOOST_REQUIRE_NO_THROW(result = static_cast<int>(a("a,b,c") * b("c,b,a")) );
   BOOST_REQUIRE_NO_THROW(result += a("a,b,c") * b("c,b,a") );
   BOOST_REQUIRE_NO_THROW(result -= a("a,b,c") * b("c,b,a") );
   BOOST_REQUIRE_NO_THROW(result *= a("a,b,c") * b("c,b,a") );
@@ -1862,7 +1862,7 @@ BOOST_AUTO_TEST_CASE( dot_expr )
 {
   // Test the dot expression function
   int result = 0;
-  BOOST_REQUIRE_NO_THROW(result = a("a,b,c") * b("a,b,c") );
+  BOOST_REQUIRE_NO_THROW(result = static_cast<int>(a("a,b,c") * b("a,b,c")) );
 
   // Compute the expected value for the dot function.
   int expected = 0;
